+++ src/main/java/com/strangegrotto/wealthdraft/networth/projections/ProjNetWorthCalcResults.java
@@ -1,6 +1,7 @@
 package com.strangegrotto.wealthdraft.networth.projections;
 
 import com.fasterxml.jackson.databind.annotation.JsonDeserialize;
+import com.strangegrotto.wealthdraft.errors.ValOrGerr;
 import org.immutables.value.Value;
 
 import java.time.LocalDate;
@@ -10,5 +11,6 @@ import java.util.SortedMap;
 @Value.Immutable
 @JsonDeserialize(as = ImmutableProjNetWorthCalcResults.class)
 public interface ProjNetWorthCalcResults {
-    Map<String, SortedMap<LocalDate, Long>> projectionsNetWorth();
+    // Mapping of scenario ID -> (projected_net_worth_values OR error)
+    Map<String, ValOrGerr<SortedMap<LocalDate, Long>>> getProjNetWorths();
 }