+++ src/test/java/com/strangegrotto/wealthdraft/errors/GerrTest.java
@@ -2,11 +2,11 @@ package com.strangegrotto.wealthdraft.errors;
 
 import org.junit.Test;
 
-public class GErrorTest {
+public class GerrTest {
     @Test
     public void testToString() {
-        GError test = GError.newError("This is a test");
-        GError propagated = GError.propagate(test, "This is the outer error");
+        Gerr test = Gerr.newGerr("This is a test");
+        Gerr propagated = Gerr.propGerr(test, "This is the outer error");
         System.out.println(propagated.toString());
     }
 }